requirements.txt
tweepy>=4.14.0
python-telegram-bot>=20.0
python-dateutil>=2.8.0 